Kill all subprocesses with `node` in their command line on a timeout
Note that this is not sufficient -- the `execAsync` code in
`@typescript/get-errors/index.ts` is ignoring the subprocess exiting
with an error (which is what happens when the process is killed), and
will continue chugging along the next `buildProject` call after the
current one is killed.  What I saw happening in debug runs is that the
node process is properly killed, and a new one is immediately
started (for just the `chakra-ui` test).

To see the problem, run this code:

```ts
async function mainAsync() {
    await execAsync(processCwd, `./kill-children-of ${processPid} node`);
    await execAsync(processCwd, `sleep 1; echo 1`);
    await execAsync(processCwd, `sleep 1; exit 1`);
    await execAsync(processCwd, `sleep 1; echo 2`);
}

async function execAsync(cwd, cmdLine) {
    console.log(`executing: ${cmdLine}`);
    return new Promise((resolve, reject) =>
        cp.exec(cmdLine, { cwd }, (err, stdout, stderr) =>
            resolve({ err, stdout, stderr })));
}
```

and see that it silently ignores the error (and runs all three
processes).