Merged PR 328183: Add User Tests
Added users tests capability. The `userTests` folder is exactly the same as https://github.com/microsoft/TypeScript/tree/main/tests/cases/user where the user tests are defined.

I also took the liberty of removing the dependency on @typescript/git-utils and @typescript/install-packages and instead added them as a file on the project. I think this will allow an easier dev experience for debugging and changes.