Merged PR 329455: Added git repo comparison capabilities
Added capability to compare different git repos and branches.

I have changed the parameters a little bit. I will agree is a little messy, specially for the Azure pipeline but I didn't want to invest too much time on getting a whole cli. I'm open to suggestions to better handle the parameters.

The parameters are intended to be obtained from the typescript bot on Wes' project.